Refactor : separate styled file and remove debug console
