+++ src/hooks/useCounter.jsx
@@ -1,19 +1,21 @@
 import { useState, useEffect } from "react";
 
-const useCounter = (num, assignable, setAssignable) => {
-  // console.log(num);
+const useCounter = (initScore, assignable, setAssignable) => {
   useEffect(() => {
-    setCount(num);
+    //re-init when profile is changed
+    setCount(initScore);
     setAssignable(0);
-  }, [num]);
+  }, [initScore]);
 
-  const [count, setCount] = useState(num);
+  const [count, setCount] = useState(initScore);
+
   const incrementCounter = () => {
     if (assignable > 0) {
       setCount((prev) => prev + 1);
       setAssignable((prev) => prev - 1);
     }
   };
+
   const decrementCounter = () => {
     if (count > 0) {
       setCount((prev) => prev - 1);
