+++ src/pages/home/index.jsx
@@ -1,45 +1,29 @@
+import ReactLoading from "react-loading";
+
 import { Switch, Route } from "react-router-dom";
 import { useEffect, useState } from "react";
+import { heroesApi } from "../../utils/common";
+import { Container, HerosList } from "./index.style";
+
+import useLoading from "../../hooks/useLoading";
 
 import Profile from "./components/Profile";
 import HeroCard from "./components/HeroCard";
-import { heroesApi } from "../../utils/common";
-import useLoading from "../../hooks/useLoading";
-import ReactLoading from "react-loading";
-import styled from "styled-components";
-
-const Container = styled.main`
-  max-width: 1560px;
-  width: 100%;
-  margin: 0 auto;
-  margin: 10% 0;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-`;
-
-const HerosList = styled.section`
-  display: grid;
-  grid-template-columns: 1fr 1fr 1fr 1fr;
-  grid-column-gap: 1.4rem;
-  width: 80%;
-  @media only screen and (max-width: 1000px) {
-    grid-template-columns: 1fr 1fr;
-  }
-  @media only screen and (max-width: 640px) {
-    grid-template-columns: 1fr;
-  }
-`;
 
 const Home = () => {
   const [herosList, setHeroList] = useState([]);
   const [loading, setLoading] = useLoading();
+
   useEffect(() => {
     const fetchHerosApi = async () => {
-      const res = await fetch(heroesApi.rootUrl);
-      const data = await res.json();
-      setHeroList(data);
-      setLoading(false);
+      try {
+        const res = await fetch(heroesApi.rootUrl);
+        const data = await res.json();
+        setHeroList(data);
+        setLoading(false);
+      } catch (err) {
+        console.error(err);
+      }
     };
     fetchHerosApi();
   }, []);
@@ -47,7 +31,7 @@ const Home = () => {
   return (
     <Container>
       {loading ? (
-        <ReactLoading color=" #ececff" />
+        <ReactLoading color="#ececff" />
       ) : (
         <HerosList>
           {herosList.map((hero) => (
